Use public imports and refactor
